Move GameController to separate package
