+++ roguelike-game/src/main/java/controllers/GameController.java
@@ -1,4 +1,4 @@
-package core;
+package controllers;
 
 import lib.Logger;
 import lib.Event;
